Moved shared models in utils in shared folder
