+++ src/h5web/visualizations/heatmap/ColorBar.tsx
@@ -2,14 +2,11 @@ import React from 'react';
 import { AxisRight } from '@vx/axis';
 import { useMeasure } from 'react-use';
 import shallow from 'zustand/shallow';
-import {
-  adaptedNumTicks,
-  generateCSSLinearGradient,
-  getDataScale,
-} from './utils';
-import styles from './HeatmapVis.module.css';
+import { adaptedNumTicks } from '../shared/utils';
 import { useInterpolator } from './hooks';
 import { useHeatmapConfig } from './config';
+import styles from './HeatmapVis.module.css';
+import { getDataScale, generateCSSLinearGradient } from './utils';
 
 function ColorBar(): JSX.Element {
   const [dataDomain, customDomain, hasLogScale] = useHeatmapConfig(
