+++ src/h5web/visualizations/shared/models.ts
@@ -1,0 +1,3 @@
+export type Domain = [number, number];
+
+export type AxisOffsets = [number, number];
