+++ src/h5web/visualizations/shared/utils.ts
@@ -1,0 +1,6 @@
+import { scaleLinear } from 'd3-scale';
+
+export const adaptedNumTicks = scaleLinear()
+  .domain([300, 900])
+  .rangeRound([3, 10])
+  .clamp(true);
